Test against "golden rom".
The code is now hard-coded to check against the nestest Rom. In addition
to the rom, there's a file detailing all the instruction it calls. We're
not there yet, but once the strings match we can do string-compare to
check if our CPU output matches that of the nestest rom.